+++ gen/com/github/themartdev/intellijgleam/lang/psi/impl/GleamHolePatternImpl.java
@@ -8,41 +8,23 @@ import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
 import com.intellij.psi.util.PsiTreeUtil;
 import static com.github.themartdev.intellijgleam.lang.psi.GleamTypes.*;
-import com.intellij.extapi.psi.ASTWrapperPsiElement;
 import com.github.themartdev.intellijgleam.lang.psi.*;
 
-public class GleamAssignmentImpl extends ASTWrapperPsiElement implements GleamAssignment {
+public class GleamHolePatternImpl extends GleamPatternImpl implements GleamHolePattern {
 
-  public GleamAssignmentImpl(@NotNull ASTNode node) {
+  public GleamHolePatternImpl(@NotNull ASTNode node) {
     super(node);
   }
 
+  @Override
   public void accept(@NotNull GleamVisitor visitor) {
-    visitor.visitAssignment(this);
+    visitor.visitHolePattern(this);
   }
 
   @Override
   public void accept(@NotNull PsiElementVisitor visitor) {
     if (visitor instanceof GleamVisitor) accept((GleamVisitor)visitor);
     else super.accept(visitor);
-  }
-
-  @Override
-  @NotNull
-  public GleamExpression getExpression() {
-    return findNotNullChildByClass(GleamExpression.class);
-  }
-
-  @Override
-  @NotNull
-  public GleamPattern getPattern() {
-    return findNotNullChildByClass(GleamPattern.class);
-  }
-
-  @Override
-  @Nullable
-  public GleamTypeAnnotation getTypeAnnotation() {
-    return findChildByClass(GleamTypeAnnotation.class);
   }
 
 }