+++ gen/com/github/themartdev/intellijgleam/lang/psi/impl/GleamPatternImpl.java
@@ -11,7 +11,7 @@ import static com.github.themartdev.intellijgleam.lang.psi.GleamTypes.*;
 import com.intellij.extapi.psi.ASTWrapperPsiElement;
 import com.github.themartdev.intellijgleam.lang.psi.*;
 
-public class GleamPatternImpl extends ASTWrapperPsiElement implements GleamPattern {
+public abstract class GleamPatternImpl extends ASTWrapperPsiElement implements GleamPattern {
 
   public GleamPatternImpl(@NotNull ASTNode node) {
     super(node);
@@ -25,54 +25,6 @@ public class GleamPatternImpl extends ASTWrapperPsiElement implements GleamPatte
   public void accept(@NotNull PsiElementVisitor visitor) {
     if (visitor instanceof GleamVisitor) accept((GleamVisitor)visitor);
     else super.accept(visitor);
-  }
-
-  @Override
-  @Nullable
-  public GleamIdentifierDiscardable getIdentifierDiscardable() {
-    return findChildByClass(GleamIdentifierDiscardable.class);
-  }
-
-  @Override
-  @Nullable
-  public GleamListPattern getListPattern() {
-    return findChildByClass(GleamListPattern.class);
-  }
-
-  @Override
-  @Nullable
-  public GleamLiteralExpr getLiteralExpr() {
-    return findChildByClass(GleamLiteralExpr.class);
-  }
-
-  @Override
-  @Nullable
-  public GleamPatternBitArray getPatternBitArray() {
-    return findChildByClass(GleamPatternBitArray.class);
-  }
-
-  @Override
-  @Nullable
-  public GleamRecordPattern getRecordPattern() {
-    return findChildByClass(GleamRecordPattern.class);
-  }
-
-  @Override
-  @Nullable
-  public GleamStringPattern getStringPattern() {
-    return findChildByClass(GleamStringPattern.class);
-  }
-
-  @Override
-  @Nullable
-  public GleamTuplePattern getTuplePattern() {
-    return findChildByClass(GleamTuplePattern.class);
-  }
-
-  @Override
-  @Nullable
-  public PsiElement getIdentifier() {
-    return findChildByType(IDENTIFIER);
   }
 
 }